requirements.txt
torch
pillow
transformers
datasets